pyproject.toml
[tool.poetry]
name = "forum"
version = "0.1.0"
description = "Platform for thematic discussions"
authors = ["Andrei Pavlov <andrio.pavlov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1"
djangorestframework = "^3.14"
psycopg2-binary = "^2.9"
gunicorn = "^20.1"
python-decouple = "^3.6"
dump-env = "^1.4"
django-split-settings = "^1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"